fix #25 fixed error and implemented themes
